requirements.txt
fastapi
uvicorn
python-multipart
pandas
google-generativeai
plotly
python-dotenv
sqlalchemy
scikit-learn
bcrypt
alembic
psycopg2-binary
